Move ResourceType and json from blobstore to physical.blobstore
