+++ datasource/src/main/scala/quasar/blobstore/azure/config.scala
@@ -17,7 +17,6 @@
 package quasar.blobstore.azure
 
 import slamdata.Predef._
-import quasar.blobstore.ResourceType
 
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.auto._
@@ -37,9 +36,15 @@ object MaxQueueSize {
   def default: MaxQueueSize = MaxQueueSize(10)
 }
 
-final case class AzureConfig(
-    containerName: ContainerName,
-    credentials: Option[AzureCredentials],
-    storageUrl: StorageUrl,
-    maxQueueSize: Option[MaxQueueSize],
-    resourceType: ResourceType)
+trait Config {
+  def containerName: ContainerName
+  def credentials: Option[AzureCredentials]
+  def storageUrl: StorageUrl
+  def maxQueueSize: Option[MaxQueueSize]
+}
+
+final case class DefaultConfig(
+    override val containerName: ContainerName,
+    override val credentials: Option[AzureCredentials],
+    override val storageUrl: StorageUrl,
+    override val maxQueueSize: Option[MaxQueueSize]) extends Config
