+++ datasource/src/main/scala/quasar/physical/blobstore/ResourceType.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.blobstore
+package quasar.physical.blobstore
 
 import slamdata.Predef.{Product, Serializable, String}
